// src/models/edit-product-values.ts
export interface EditProductValues {
    image: string;
    name: string;
    amount: number;
    fullPrise: number;
    count: number;
    description?: string;
    isDiscount: boolean;
    isNovelty: boolean;
    brand: string;
    gender: string;
    perfumeType: string;
    fragrance: string[];
}